Moved TGA Header methods back into header.rs
