+++ src/tga/encoder.rs
@@ -1,11 +1,6 @@
-use std::{error, fmt, io::{self, Write}};
-
-use super::header::{Header, ImageType, ALPHA_BIT_MASK, SCREEN_ORIGIN_BIT_MASK};
-use crate::color::ColorType;
-use crate::error::{EncodingError, ImageResult, UnsupportedError, UnsupportedErrorKind};
-use crate::{image::ImageEncoder, ImageError, ImageFormat};
-use byteorder::{LittleEndian, WriteBytesExt};
-use std::convert::TryFrom;
+use super::header::Header;
+use crate::{error::EncodingError, ColorType, ImageEncoder, ImageError, ImageFormat, ImageResult};
+use std::{convert::TryFrom, error, fmt, io::Write};
 
 /// Errors that can occur during encoding and saving of a TGA image.
 #[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
@@ -59,6 +54,13 @@ impl<W: Write> TgaEncoder<W> {
         height: u32,
         color_type: ColorType,
     ) -> ImageResult<()> {
+        // Validate dimensions.
+        let width = u16::try_from(width)
+            .map_err(|_| ImageError::from(EncoderError::WidthInvalid(width)))?;
+
+        let height = u16::try_from(height)
+            .map_err(|_| ImageError::from(EncoderError::HeightInvalid(height)))?;
+
         // Write out TGA header.
         let header = Header::from_pixel_info(color_type, width, height)?;
         header.write_to(&mut self.writer)?;
@@ -92,75 +94,11 @@ impl<W: Write> ImageEncoder for TgaEncoder<W> {
     }
 }
 
-impl Header {
-    /// Load the header with values from pixel information.
-    pub(crate) fn from_pixel_info(
-        color_type: ColorType,
-        width: u32,
-        height: u32,
-    ) -> ImageResult<Self> {
-        let mut header = Self::default();
-
-        if width > 0 && height > 0 {
-            header.image_width = u16::try_from(width)
-                .map_err(|_| ImageError::from(EncoderError::WidthInvalid(width)))?;
-
-            header.image_height = u16::try_from(height)
-                .map_err(|_| ImageError::from(EncoderError::HeightInvalid(height)))?;
-
-            let (num_alpha_bits, other_channel_bits, image_type) = match color_type {
-                ColorType::Rgba8 | ColorType::Bgra8 => (8, 24, ImageType::RawTrueColor),
-                ColorType::Rgb8 | ColorType::Bgr8 => (0, 24, ImageType::RawTrueColor),
-                ColorType::La8 => (8, 8, ImageType::RawGrayScale),
-                ColorType::L8 => (0, 8, ImageType::RawGrayScale),
-                _ => {
-                    return Err(ImageError::Unsupported(
-                        UnsupportedError::from_format_and_kind(
-                            ImageFormat::Tga.into(),
-                            UnsupportedErrorKind::Color(color_type.into()),
-                        ),
-                    ))
-                }
-            };
-
-            header.image_type = image_type as u8;
-            header.pixel_depth = num_alpha_bits + other_channel_bits;
-            header.image_desc = num_alpha_bits & ALPHA_BIT_MASK;
-            header.image_desc |= SCREEN_ORIGIN_BIT_MASK; // Upper left origin.
-        }
-
-        Ok(header)
-    }
-
-    /// Write out the header values.
-    pub(crate) fn write_to(&self, w: &mut dyn Write) -> io::Result<()> {
-        w.write_u8(self.id_length)?;
-        w.write_u8(self.map_type)?;
-        w.write_u8(self.image_type)?;
-        w.write_u16::<LittleEndian>(self.map_origin)?;
-        w.write_u16::<LittleEndian>(self.map_length)?;
-        w.write_u8(self.map_entry_size)?;
-        w.write_u16::<LittleEndian>(self.x_origin)?;
-        w.write_u16::<LittleEndian>(self.y_origin)?;
-        w.write_u16::<LittleEndian>(self.image_width)?;
-        w.write_u16::<LittleEndian>(self.image_height)?;
-        w.write_u8(self.pixel_depth)?;
-        w.write_u8(self.image_desc)?;
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::{EncoderError, TgaEncoder};
-    use crate::tga::TgaDecoder;
-
-    use crate::color::ColorType;
-    use crate::image::ImageDecoder;
-    use crate::ImageError;
-
-    use std::error::Error;
-    use std::io::Cursor;
+    use crate::{tga::TgaDecoder, ColorType, ImageDecoder, ImageError};
+    use std::{error::Error, io::Cursor};
 
     fn round_trip_image(image: &[u8], width: u32, height: u32, c: ColorType) -> Vec<u8> {
         let mut encoded_data = Vec::new();
